Refactor Ground Clue functions
