+++ src/main/java/com/cluedetails/ClueGroundManager.java
@@ -24,6 +24,7 @@
  */
 package com.cluedetails;
 
+import com.cluedetails.filters.ClueTier;
 import java.util.stream.Collectors;
 import lombok.Getter;
 import net.runelite.api.*;
@@ -266,7 +267,8 @@ public class ClueGroundManager
 
 	private void removeDespawnedClues()
 	{
-		groundClues.entrySet().removeIf(entry -> {
+		groundClues.entrySet().removeIf(entry ->
+		{
 			List<ClueInstance> cluesList = entry.getValue();
 			cluesList.removeIf(clueInstance -> clueInstance.getDespawnTick(client.getTickCount()) <= client.getTickCount());
 			return cluesList.isEmpty();
@@ -366,7 +368,8 @@ public class ClueGroundManager
 			.map(tileItem -> sortedStoredClues.stream()
 				.filter(clue -> clue.getTileItem() == tileItem)
 				.findFirst()
-				.orElseGet(() -> {
+				.orElseGet(() ->
+				{
 					ClueInstance clueInstance = new ClueInstance(List.of(), tileItem.getId(), tileWp, tileItem, client.getTickCount());
 					clueInstance.setTileItem(tileItem);
 					return clueInstance;
@@ -427,6 +430,108 @@ public class ClueGroundManager
 		return items.stream()
 			.filter(item -> Clues.isClue(item.getId(), clueDetailsPlugin.isDeveloperMode()))
 			.collect(Collectors.toList());
+	}
+
+	class ClueInstanceComparator implements Comparator<ClueInstance>
+	{
+		@Override
+		public int compare(ClueInstance o1, ClueInstance o2)
+		{
+			// Primary comparison: Compare by despawn time
+			int despawnComparison = Integer.compare(o1.getTicksToDespawnConsideringTileItem(client.getTickCount()),
+				o2.getTicksToDespawnConsideringTileItem(client.getTickCount()));
+			if (despawnComparison != 0)
+			{
+				return despawnComparison;
+			}
+			else
+			{
+				// Secondary comparison: If despawn times are the same, compare by itemId
+				return Integer.compare(o1.getItemId(), o2.getItemId());
+			}
+		}
+	}
+
+	public TreeMap<ClueInstance, Integer> getClueInstancesWithQuantityAtWp(ClueDetailsConfig config, WorldPoint wp, int currentTick)
+	{
+		if (groundClues.get(wp).isEmpty()) return null;
+
+		List<ClueInstance> groundItemList = groundClues.get(wp);
+		Map<ClueInstance, Integer> groundItemMap = new HashMap<>();
+
+		if (config.collapseGroundCluesByTier())
+		{
+			groundItemMap = keepOldestTierClues(groundItemList, currentTick);
+		}
+		else if (config.collapseGroundClues())
+		{
+			groundItemMap = keepOldestUniqueClues(groundItemList, currentTick);
+		}
+		else
+		{
+			for (ClueInstance item : groundItemList)
+			{
+				groundItemMap.put(item, 1);
+			}
+		}
+
+		// Sort ClueInstances by despawn time
+		ClueInstanceComparator clueInstanceComparator = new ClueInstanceComparator();
+		TreeMap<ClueInstance, Integer> clueInstancesWithQuantityAtWp = new TreeMap<>(clueInstanceComparator);
+		clueInstancesWithQuantityAtWp.putAll(groundItemMap);
+		return clueInstancesWithQuantityAtWp;
+	}
+
+	// Remove duplicate step clues, maintaining a count of the original amount of each
+	public static Map<ClueInstance, Integer> keepOldestUniqueClues(List<ClueInstance> items, int currentTick)
+	{
+		Map<List<Integer>, ClueInstance> lowestValueItems = new HashMap<>();
+		Map<List<Integer>, Integer> uniqueCount = new HashMap<>();
+
+		for (ClueInstance item : items)
+		{
+			List<Integer> clueIds = item.getClueIds();
+
+			if (!lowestValueItems.containsKey(clueIds)
+				|| item.getDespawnTick(currentTick) < lowestValueItems.get(clueIds).getDespawnTick(currentTick))
+			{
+				lowestValueItems.put(clueIds, item);
+				uniqueCount.put(clueIds, 1);
+			}
+			else
+			{
+				uniqueCount.put(clueIds, uniqueCount.get(clueIds) + 1);
+			}
+		}
+
+		return lowestValueItems.values().stream()
+			.collect(Collectors.toMap(item -> item, item -> uniqueCount.get(item.getClueIds())));
+	}
+
+	// Remove duplicate tier clues, maintaining a count of the original amount of each
+	public static Map<ClueInstance, Integer> keepOldestTierClues(List<ClueInstance> items, int currentTick)
+	{
+		Map<ClueTier, ClueInstance> lowestValueItems = new HashMap<>();
+		Map<ClueTier, Integer> uniqueCount = new HashMap<>();
+
+		for (ClueInstance item : items)
+		{
+			ClueTier tier = item.getTier();
+
+			if (!lowestValueItems.containsKey(tier)
+				|| item.getDespawnTick(currentTick) < lowestValueItems.get(tier).getDespawnTick(currentTick))
+			{
+				lowestValueItems.put(tier, item);
+				uniqueCount.put(tier, 1);
+			}
+			else
+			{
+				uniqueCount.put(tier, uniqueCount.get(tier) + 1);
+			}
+		}
+
+		return lowestValueItems.values().stream()
+			.collect(Collectors.toMap(item -> item, item ->	uniqueCount.get(item.getTier())));
 	}
 
 	public void saveStateToConfig()
