+++ src/main/java/com/cluedetails/ClueInstance.java
@@ -35,6 +35,8 @@ import net.runelite.api.ItemID;
 import net.runelite.api.TileItem;
 import net.runelite.api.coords.WorldPoint;
 import net.runelite.client.config.ConfigManager;
+import net.runelite.client.util.QuantityFormatter;
+import org.apache.commons.text.WordUtils;
 
 @Data
 public class ClueInstance
@@ -106,6 +108,77 @@ public class ClueInstance
 			return null;
 		}
 		return clue.getClueTier();
+	}
+
+	public String getGroundText(ClueDetailsPlugin plugin, ClueDetailsConfig config, ConfigManager configManager, int quantity)
+	{
+		StringBuilder itemStringBuilder = new StringBuilder();
+		List<Integer> clueIds = this.getClueIds();
+
+		if (clueIds.isEmpty())
+		{
+			itemStringBuilder.append(this.getItemName(plugin));
+		}
+		else
+		{
+			int clueId = this.getClueIds().get(0);
+			Clues clueDetails = Clues.forClueIdFiltered(clueId);
+
+			if (clueDetails == null)
+			{
+				return null;
+			}
+
+			String clueText;
+			if (config.changeGroundClueText() && !config.collapseGroundCluesByTier())
+			{
+				if (clueIds.size() > 1)
+				{
+					clueText = "Three-step (master)";
+				}
+				else
+				{
+					clueText = clueDetails.getDetail(configManager);
+				}
+			}
+			else
+			{
+				clueText = WordUtils.capitalizeFully(clueDetails.getClueTier().toString().replace("_", " "));
+			}
+
+			itemStringBuilder.append(clueText);
+		}
+
+		if ((config.collapseGroundClues() || config.collapseGroundCluesByTier()) && quantity > 1)
+		{
+			itemStringBuilder.append(" (")
+				.append(QuantityFormatter.quantityToStackSize(quantity))
+				.append(')');
+		}
+		return itemStringBuilder.toString();
+	}
+
+	public Color getGroundColor(ClueDetailsConfig config, ConfigManager configManager)
+	{
+		Color color = Color.WHITE;
+		List<Integer> clueIds = this.getClueIds();
+
+		if (!clueIds.isEmpty())
+		{
+			int clueId = this.getClueIds().get(0);
+			Clues clueDetails = Clues.forClueIdFiltered(clueId);
+
+			if (clueDetails == null)
+			{
+				return color;
+			}
+
+			if (config.colorGroundClues())
+			{
+				color = clueDetails.getDetailColor(configManager);
+			}
+		}
+		return color;
 	}
 
 	public int getDespawnTick(int currentTick)
@@ -194,6 +267,6 @@ public class ClueInstance
 		{
 			return config.masterDetails();
 		}
-		return true;
+		else return getTier() != null;
 	}
 }
